Big update: increased FFT length + output data reordering
-New limits are: pow(2,30) in all dimensions in C2C, pow(2,30) y and z in R2C C2R (non-unit x). Done by implementing three pass four-step FFT. If reordering is not performed - max sequence length is pow(2,32). It will be harder to go further, as pow(2,32) is a limit that can be addressed by a 32-bit uint
-Added an option to return data to the original layout and unshuffle four-step FFT results (will produce the same result as cuFFT and FFTW). Uses additional buffer(allocated by user) to do transposition without additional memory uploads
-Convolutions can be done without additional buffer and transposition (save in memory)
-Made a custom dispatch scheme, that accounts to 16bit addressing limits, increasing it to 32-bit - this allows to do big FFTs in y and z directions
-Updated performance scripts to cover full supported range (~60 systems) in all precisions. Results attached

+++ lib/vkFFT.h
@@ -3188,8 +3188,8 @@ extern "C" {
 
 	}
 	void dispatchEnchanced(VkFFTApplication* app, VkCommandBuffer commandBuffer, VkFFTAxis* axis, uint32_t* dispatchBlock) {
-		uint32_t maxBlockPow2Size[3] = { pow(2,(uint32_t)log2(app->configuration.maxComputeWorkGroupCount[0])),pow(2,(uint32_t)log2(app->configuration.maxComputeWorkGroupCount[1])),pow(2,(uint32_t)log2(app->configuration.maxComputeWorkGroupCount[2])) };
-		uint32_t blockNumber[3] = { ceil(dispatchBlock[0]/ (float)maxBlockPow2Size[0] ),ceil(dispatchBlock[1] / (float)maxBlockPow2Size[1]),ceil(dispatchBlock[2] / (float)maxBlockPow2Size[2]) };
+		uint32_t maxBlockPow2Size[3] = { (uint32_t)pow(2,(uint32_t)log2(app->configuration.maxComputeWorkGroupCount[0])),(uint32_t)pow(2,(uint32_t)log2(app->configuration.maxComputeWorkGroupCount[1])),(uint32_t)pow(2,(uint32_t)log2(app->configuration.maxComputeWorkGroupCount[2])) };
+		uint32_t blockNumber[3] = { (uint32_t)ceil(dispatchBlock[0]/ (float)maxBlockPow2Size[0] ),(uint32_t)ceil(dispatchBlock[1] / (float)maxBlockPow2Size[1]),(uint32_t)ceil(dispatchBlock[2] / (float)maxBlockPow2Size[2]) };
 		for (uint32_t i = 0; i < 3; i++)
 			if (blockNumber[i] == 1) maxBlockPow2Size[i] = dispatchBlock[i];
 		for (uint32_t i = 0; i < blockNumber[0]; i++) {
